set up for local theme development
Installing a node package from local file breaks npm "peerDependencies".
This is because the local install is always going to point at the "node_modules" folder above the child plugin.
This bereaks typedoc because typedoc is loaded twice.

Instead, we develope and build the plugin/theme in the local directory and explicity call the local build artifact from typedoc.js